#294 Extract AddMemberDialog
