+++ client/app/routes/dashboard/settings/organization/members.tsx
@@ -3,16 +3,12 @@ import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { useTRPC } from "~/utils/trpc/react";
 import { useOrganizationContext } from "~/contexts/OrganizationContext";
 import Skeleton from "~/components/atoms/skeleton/Skeleton";
-import { useForm } from "react-hook-form";
 import Button from "~/components/atoms/button/Button";
 import { Dialog, DialogContent } from "~/components/atoms/dialog/Dialog";
 import ErrorBanner from "~/components/atoms/banner/ErrorBanner";
 import { AvatarIcon } from "~/components/atoms/avatar/Avatar";
 import Breadcrumbs from "~/components/organisms/breadcrumbs/Breadcrumbs";
-
-interface FormInput {
-  email: string;
-}
+import AddMemberDialog from "~/components/organisms/dialog/AddMemberDialog";
 
 interface MemberToRemove {
   userId: number;
@@ -36,35 +32,9 @@ export default function Members() {
     }),
   );
 
-  const sendInviteMutation = useMutation(
-    trpc.organization.sendInvite.mutationOptions(),
-  );
-
   const removeMemberMutation = useMutation(
     trpc.organization.removeOrganizationMember.mutationOptions(),
   );
-
-  const { register, handleSubmit, reset, watch } = useForm<FormInput>({
-    defaultValues: { email: "" },
-  });
-
-  const emailInput = watch("email", "");
-
-  function onInviteMember(data: FormInput) {
-    sendInviteMutation.mutate(
-      {
-        organizationId: organization.id,
-        toEmail: data.email,
-        inviteUrlPrefix: `${window.location.origin}/organizations/invite/`,
-      },
-      {
-        onSuccess: () => {
-          reset();
-          setShowAddMemberDialog(false);
-        },
-      },
-    );
-  }
 
   function openRemoveMemberDialog(member: MemberToRemove) {
     removeMemberMutation.reset();
@@ -218,54 +188,11 @@ export default function Members() {
       </div>
 
       {organization.isOwner && (
-        <Dialog
+        <AddMemberDialog
+          organizationId={organization.id}
           open={showAddMemberDialog}
           onOpenChange={setShowAddMemberDialog}
-        >
-          <DialogContent>
-            <div className="flex flex-col gap-4">
-              <div className="flex flex-col gap-2">
-                <h1>Invite Member</h1>
-                <p className="text-mauve-11 text-sm">
-                  Invite members via email. They&apos;ll receive a link to join
-                  your organization.
-                </p>
-              </div>
-              <form
-                className="flex flex-col gap-3"
-                onSubmit={handleSubmit(onInviteMember)}
-              >
-                <input
-                  type="email"
-                  className="border-mauve-6 text-mauve-12 placeholder:text-mauve-11 bg-gray-2 w-full rounded-md border p-2 text-sm shadow-[inset_0_1px_2px_rgba(0,0,0,0.12)]"
-                  placeholder="Email*"
-                  {...register("email")}
-                />
-                <div className="flex justify-end gap-2">
-                  <Button
-                    type="button"
-                    intent="secondary"
-                    className="w-24"
-                    onClick={() => {
-                      setShowAddMemberDialog(false);
-                      reset();
-                      sendInviteMutation.reset();
-                    }}
-                  >
-                    Cancel
-                  </Button>
-                  <Button
-                    className="h-10 w-24"
-                    type="submit"
-                    disabled={!emailInput || sendInviteMutation.isPending}
-                  >
-                    Invite
-                  </Button>
-                </div>
-              </form>
-            </div>
-          </DialogContent>
-        </Dialog>
+        />
       )}
 
       <Dialog
